+++ examples/express-postgres/CHANGELOG.md
@@ -1,5 +1,17 @@
 # express-postgres
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @meshql/core@0.7.0
+  - @meshql/client@0.5.3
+  - @meshql/http@0.5.3
+  - @meshql/integrity@0.5.3
+  - @meshql/postgres@0.5.3
+  - @meshql/upload@0.5.3
+
 ## 0.1.6
 
 ### Patch Changes